Moving extraction of station staff details from _stations.py to _data_parsing.py as this is a derived output
